abstract base miner
