+++ openminers/__init__.py
@@ -1,4 +1,5 @@
 # Base Miner imports
+from .base.miner import BaseMiner as BaseMiner
 from .base.prompting_miner import BasePromptingMiner as BasePromptingMiner
 from .base.embedding_miner import BaseEmbeddingMiner as BaseEmbeddingMiner
