+++ openminers/base/prompting_miner.py
@@ -15,84 +15,47 @@
 # OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 # DEALINGS IN THE SOFTWARE.
 
-import copy
-import wandb
 import torch
 import argparse
-import threading
 import bittensor as bt
 
-from abc import ABC, abstractmethod
+from abc import ABC
 from typing import List, Dict, Union, Tuple
 
-from .run import run
 from .forward import forward
 from .priority import priority
 from .blacklist import blacklist
-from .mock import MockSubtensor
-from .config import config, check_config
+from .miner import BaseMiner
 
-class BasePromptingMiner( ABC ):
 
-    @classmethod
-    def config( cls ) -> "bt.Config": return config( cls )
+class BasePromptingMiner( BaseMiner, ABC ):
 
     @classmethod
-    @abstractmethod
-    def add_args( cls, parser: argparse.ArgumentParser ):
-        ...
+    def config( cls ) -> "bt.Config":
+        parser = argparse.ArgumentParser()
+        cls.add_super_args( parser )
+        return bt.config( parser )
 
-    @abstractmethod
-    def forward( self, messages: List[Dict[str, str]] ) -> str:
-        ...
-
-    def priority( self, forward_call: "bt.TextPromptingForwardCall" ) -> float:
-        raise NotImplementedError('priority not implemented in subclass')
-    
-    def blacklist( self, forward_call: "bt.TextPromptingForwardCall" ) -> Union[ Tuple[bool, str], bool ]:
-        raise NotImplementedError('blacklist not implemented in subclass')
-
-    def __init__( 
-            self, 
-            config: "bt.Config" = None,
-            axon: "bt.axon" = None,
-            wallet: "bt.Wallet" = None,
-            subtensor: "bt.Subtensor" = None,
-            synapse: "bt.Synapse" = None,
-        ):
-
-        # Instantiate and check configs.
-        # Grab super config.
-        super_config = copy.deepcopy( config or BasePromptingMiner.config() )
-
-        # Grab child config
-        self.config = self.config()
-
-        # Merge them, but overwrite from the child config.
-        self.config.merge( super_config )
-        check_config( BasePromptingMiner, self.config )
-
-        # Instantiate logging.
-        bt.logging( config = self.config, logging_dir = self.config.miner.full_path )
-
-        # Instantiate subtensor.
-        if self.config.miner.mock_subtensor:
-            self.subtensor = subtensor or MockSubtensor( self.config )
-        else:
-            self.subtensor = subtensor or bt.subtensor( self.config )
-
-        # Instantiate metagraph.
-        self.metagraph = self.subtensor.metagraph( self.config.netuid )
-
-        # Instantiate wallet.
-        self.wallet = wallet or bt.wallet( self.config )
-
-        # Instantiate axon.
-        self.axon = axon or bt.axon(
-            wallet = self.wallet,
-            metagraph = self.metagraph,
-            config = self.config,
+    @classmethod
+    def add_super_args( cls, parser: argparse.ArgumentParser ):
+        """ Add arguments specific to BasePromptingMiner to parser.
+        """
+        cls.add_args(parser)
+        parser.add_argument(
+            '--neuron.max_batch_size', 
+            type = int, 
+            help = 'The maximum batch size for forward requests.',
+            default = -1
         )
+        parser.add_argument(
+            '--neuron.max_sequence_len', 
+            type = int, 
+            help = 'The maximum sequence length for forward requests.',
+            default = -1
+        )
+
+    def __init__( self, *args, **kwargs ):
+        super( BasePromptingMiner, self ).__init__( *args, **kwargs )
 
         # Define synapse.
         class Synapse( bt.TextPromptingSynapse ):
@@ -100,7 +63,7 @@ class BasePromptingMiner( ABC ):
             # Build priority function.
             def priority( _, forward_call: "bt.TextPromptingForwardCall" ) -> float:
                 return priority( self, self.priority, forward_call )
-            
+
             # Build blacklist function.
             def blacklist( _, forward_call: "bt.TextPromptingForwardCall" ) -> Union[ Tuple[bool, str], bool ]:
                 return blacklist( self, self.blacklist, forward_call )
@@ -113,44 +76,6 @@ class BasePromptingMiner( ABC ):
             # TODO(const): accept this.
             def backward( self, messages: List[Dict[str, str]], response: str, rewards: torch.FloatTensor ) -> str: 
                 pass
-                  
+
         # Instantiate synapse.
-        self.synapse = synapse or Synapse( axon = self.axon )
-
-        # Init wandb.
-        if self.config.wandb.on:
-            wandb.init(
-                project = self.config.wandb.project_name,
-                entity = self.config.wandb.entity,
-                config = self.config,
-                mode = 'offline' if self.config.wandb.offline else 'online',
-                dir = self.config.miner.full_path,
-                magic = True,
-            )
-
-        # Instantiate runners.
-        self.should_exit: bool = False
-        self.is_running: bool = False
-        self.thread: threading.Thread = None
-
-    def run( self ): run( self )
-
-    def run_in_background_thread(self):
-        if not self.is_running:
-            bt.logging.debug( f'Starting miner background thread') 
-            self.should_exit = False
-            self.thread = threading.Thread( target = self.run, daemon = True )
-            self.thread.start()
-            self.is_running = True
-            bt.logging.debug( f'Started') 
-
-    def stop_run_thread(self):
-        if self.is_running:
-            bt.logging.debug( f'Stopping miner background thread...') 
-            self.should_exit = True
-            self.thread.join( 5 )
-            bt.logging.debug( f'Stopped') 
-
-    def __enter__( self ): self.run_in_background_thread()
-
-    def __exit__( self, exc_type, exc_value, traceback ): self.stop_run_thread()
+        self.synapse = Synapse( axon = self.axon )
